Coach navigation and mapping logic updated
- Implemented `CoachEdit` route in `Screen.kt` and added its destination to `TrainMonitoringFlow.kt`.
- Updated `MonitoringTrainInProgress` to handle coach addition and editing navigation via `CoachSelectionDialog`.
- Added mapping logic for converting coach UI models to domain and draft states in `PassengerCoachMappers.kt`.
- Enhanced `TrainOrderViewModel` with `getOrder()` and improved crew management logic (ID handling and mapping).
- Added `fromString` utility to `CoachTypes` enum for easier type resolution.
- Updated `AddWorkerDialog` and `WorkerUI` usage to standardize tab number handling.
- Integrated `EditCoachEntryPoint` into the navigation graph for managing coach edits.

+++ app/src/main/java/com/adrzdv/mtocp/ui/component/dialogs/AddWorkerDialog.kt
@@ -67,7 +67,7 @@ fun AddWorkerDialog(
                         selectedPosition?.let { position ->
                             onWorkerAdd(
                                 WorkerUI(
-                                    tabNumber.toInt(),
+                                    tabNumber,
                                     name,
                                     position.description,
                                     selectedDepot
